Changed parameters in ai.yaml for better centering. Special thanks to Nitthilan for helping finding issues with center state, and parameters Clean up of old AI files.
