+++ param/ai.yaml
@@ -22,10 +22,10 @@ ai:
     search_gate:
         yaw_speed_factor: 2.0 # how fast to yaw while searching gate
     center:
-        errorGoal: 0.1 # what is the error Goal, in precentage so 0.1 is 10%
+        errorGoal: 0.02 # what is the error Goal, in precentage so 0.1 is 10%
         yaw_factor: -2.5 # yaw factor for centering. Negative is to the right
         dive_factor: -2.0 # dive factor for centering. Negative is to dive
     move_forward:
-        errorGoal: 0.01 #What is the error Goal in moving forward
+        errorGoal: 0.03 #What is the error Goal in moving forward
         distanceGoal: 0.01 # What is the distanceGoal, teremined from testing
         forward_speed: 0.4 # What is the speed with which we need to approach
