--- param/ai_manager.yaml
@@ -1,31 +1,0 @@
-tasks:
-  -  name: "Forward_Gate"
-     node: "forward_until.py"
-     args: "-45 start_gate_post"
-     remap_args: []
-     timeout_secs: 20
-  -  name: "Gate"
-     node: "gate_task.py"
-     args: "-d 10"
-     remap_args: []
-     timeout_secs: 60
-  -  name: "Forward_Buoys"
-     node: "forward_until.py"
-     args: "-10 red_buoy yellow_buoy green_bouy"
-     remap_args: []
-     timeout_secs: 20
-  -  name: "Red Buoy"
-     node: "buoy_task.py"
-     args: "red_buoy"
-     remap_args: []
-     timeout_secs: 30
-  -  name: "Green Buoy"
-     node: "buoy_task.py"
-     args: "green_buoy"
-     remap_args: []
-     timeout_secs: 30
-  -  name: "Yellow Buoy"
-     node: "buoy_task.py"
-     args: "yellow_buoy"
-     remap_args: []
-     timeout_secs: 30
