+++ src/ai/util_states.py
@@ -1,0 +1,224 @@
+#!/usr/bin/python
+import rospy
+from util import *
+from rs_yolo.msg import DetectionArray
+from robosub.msg import control
+from SubscribeState import SubscribeState
+from control_wrapper import control_wrapper
+import smach
+import smach_ros
+
+search_direction = 'right'
+# Based on move_forward from forward_until.py with tweaks for vision
+class move_to_gate(SubscribeState):
+    def __init__(self, vision_label, poll_rate=10):
+        SubscribeState.__init__(self, "vision", DetectionArray, self.callback,
+                               outcomes=['success'])
+        self.vision_label = vision_label
+        self._poll_rate = rospy.Rate(poll_rate)
+        self.forward_error = rospy.get_param("ai/move_to_gate/forward_error")
+
+    def callback(self, detectionArray):
+        c = control_wrapper()
+        c.levelOut()
+        # forward_error is a parameter
+        c.forwardError(self.forward_error)
+
+        detections = filterByLabel(detectionArray.detections,
+            self.vision_label)
+        vision_result = getNMostProbable(detections, 2, thresh=0.5)
+        normalize(vision_result)
+        # move forward until we see a gate task post
+        if len(vision_result) == 0:
+            c.publish()
+            self._poll_rate.sleep()
+        self.exit("success")
+        return 'success'
+
+# Lost state for when AI feels lost
+# When we do not see any of gate posts at all
+class lost(SubscribeState):
+    def __init__(self, vision_label, poll_rate=10):
+        SubscribeState.__init__(self, "vision", DetectionArray, self.callback,
+                               outcomes=['1 post', 'none', '2 posts'])
+        self.vision_label = vision_label
+        self.yaw = rospy.get_param("ai/search_gate/yaw_speed_factor")
+        self._poll_rate = rospy.Rate(poll_rate)
+
+    def callback(self, detectionArray):
+        global search_direction
+        c = control_wrapper()
+        c.forwardError(0.0)
+        detections = filterByLabel(detectionArray.detections,
+            self.vision_label)
+
+        vision_result = getNMostProbable(detections, 2, thresh=0.5)
+
+        currentYaw = 0
+        if search_direction == 'right':
+            currentYaw = -self.yaw
+        else:
+            currentYaw = self.yaw
+        c.yawLeftRelative(currentYaw)
+
+        if len(vision_result) == 0:
+            rospy.loginfo("search direction {}".format(search_direction))
+            rospy.loginfo("yaw: {}".format(currentYaw))
+            c.publish()
+            self._poll_rate.sleep()
+            self.exit('none')
+        elif len(vision_result) == 1:
+            self.exit('1 post')
+        elif len(vision_result) == 2:
+            self.exit('2 posts')
+
+# Flips direction of search
+class flip(smach.State):
+    def __init__(self):
+        smach.State.__init__(self, outcomes=['success'])
+
+    def execute(self, userdata):
+        global search_direction
+        if search_direction == 'right':
+            search_direction = 'left'
+        else:
+            search_direction = 'right'
+        return 'success'
+
+# High level state machine for searching gate posts
+class Search_for_gates(smach.StateMachine):
+    def __init__(self):
+        smach.StateMachine.__init__(self, outcomes=['success'])
+
+        with self:
+            smach.StateMachine.add("LOST", lost('gate_post'),
+                transitions={'1 post': 'SEARCH', 'none': 'LOST',
+                '2 posts': 'SEARCH'})
+
+            smach.StateMachine.add("FLIP", flip(),
+                                  transitions={'success': 'LOST'})
+
+            smach.StateMachine.add("SEARCH", search('gate_post'),
+                transitions={'2 posts': 'success', 'none': 'FLIP',
+                '1 post': 'SEARCH'})
+
+# State for searching gate posts. Comes after state Lost, so it searches,
+# and yaws same direction as it was in Lost state
+# By default yaws to the right until
+# it either loses all of posts or finds second.
+class search(SubscribeState):
+    def __init__(self, vision_label, poll_rate=10):
+        SubscribeState.__init__(self, "vision", DetectionArray, self.callback,
+                               outcomes=['2 posts', 'none', '1 post'])
+        self.vision_label = vision_label
+        self.yaw = rospy.get_param("ai/search_gate/yaw_speed_factor")
+        self._poll_rate = rospy.Rate(poll_rate)
+
+    def callback(self, detectionArray):
+        global search_direction
+        c = control_wrapper()
+        c.forwardError(0.0)
+
+        detections = filterByLabel(detectionArray.detections,
+            self.vision_label)
+        vision_result = getNMostProbable(detections, 2, thresh=0.5)
+
+        rospy.loginfo("results: {}".format(len(vision_result)))
+
+        currentYaw = 0
+        if search_direction == 'right':
+            currentYaw = -self.yaw
+        else:
+            currentYaw = self.yaw
+
+        c.yawLeftRelative(currentYaw)
+        if len(vision_result) == 1:
+            c.publish()
+            self._poll_rate.sleep()
+            self.exit('1 post')
+        elif len(vision_result) == 2:
+            self.exit('2 posts')
+        else:
+            rospy.loginfo("search direction {}".format(search_direction))
+            self.exit('none')
+
+# State for centering between gate posts or moving while being centered
+class center(SubscribeState):
+    def __init__(self, vision_label):
+        SubscribeState.__init__(self, "vision", DetectionArray, self.callback,
+                               outcomes=['centered', 'lost'])
+        self.vision_label = vision_label
+        self.errorGoal = rospy.get_param("ai/center/errorGoal")
+        self.yaw_factor = rospy.get_param("ai/center/yaw_factor")
+        self.dive_factor = rospy.get_param("ai/center/dive_factor")
+
+
+    def callback(self, detectionArray):
+        c = control_wrapper()
+        c.levelOut()
+
+        detections = filterByLabel(detectionArray.detections,
+            self.vision_label)
+        vision_result = getNMostProbable(detections, 2, thresh=0.5)
+
+        if len(vision_result) < 2:
+            self.exit('lost')
+            return 'lost'
+
+        gateXPos = (vision_result[0].x + vision_result[1].x) / 2
+        gateYPos = (vision_result[0].y + vision_result[1].y) / 2
+
+        rospy.loginfo(("Gate X: {}\tGate Y: {}".format(gateXPos, gateYPos)))
+
+        if abs(gateXPos-0.5) > self.errorGoal:
+            # If we are not centered by yaw
+            yaw_left = (gateXPos-0.5) * self.yaw_factor
+            c.yawLeftRelative(yaw_left * 60)
+            rospy.loginfo("trying to yaw: {}".format(yaw_left * 60))
+        elif abs(gateYPos-0.5) > self.errorGoal:
+            # If our depth is not enough for centering
+            dive = (gateYPos-0.5) * self.dive_factor
+            rospy.loginfo("trying to dive: {}".format(dive))
+            c.diveRelative(dive)
+        else:
+            self.exit('centered')
+
+        c.publish()
+
+# State for moving forward while checking if we are centered
+# In future need to make this function to watch for bottom facing camera,
+# slow  down or stop if it gets close to path marker.
+class move_forward_centered(SubscribeState):
+    def __init__(self, vision_label):
+        SubscribeState.__init__(self, "vision", DetectionArray, self.callback,
+                               outcomes=['ready', 'not centered', 'lost'])
+        self.vision_label = vision_label
+        self.distanceGoal = rospy.get_param("ai/move_forward/distanceGoal")
+        self.errorGoal = rospy.get_param("ai/move_forward/errorGoal")
+        self.forward_speed = rospy.get_param("ai/move_forward/forward_speed")
+
+    def callback(self, detectionArray):
+        c = control_wrapper()
+        c.levelOut()
+        c.forwardError(self.forward_speed)
+
+        detections = filterByLabel(detectionArray.detections,
+            self.vision_label)
+        vision_result = getNMostProbable(detections, 2, thresh=0.5)
+
+        gateXPos = (vision_result[0].x + vision_result[1].x) / 2
+        gateYPos = (vision_result[0].y + vision_result[1].y) / 2
+
+        if (((vision_result[0].width * vision_result[0].height) +
+           (vision_result[1].width *
+           vision_result[1].height)) > self.distanceGoal):
+            self.exit('ready')
+            return 'ready'
+        if abs(gateXPos-0.5) > self.errorGoal or abs(gateYPos-0.5) > self.errorGoal:
+            self.exit('not centered')
+        if len(vision_result) < 2:
+            self.exit('lost')
+        c.publish()
+
+if __name__ == "__main__":
+    pass
